foxtv: last programme on each day was missing
The HTML data does not provide the stop time for the last entry of a
day. As we operate on a weekly page, for Monday to Saturday we can
simply look at the first entry for the next day and use this as the
stop time.

So we only loose information for the last entry on Sundays.