telkku: implement channel() method
The new web site uses JavaScript but no dynamic data fetch from the
backend. The programme information is instead encoded in a single JSON
string embedded in a script node.

Channels are grouped so we fetch the URL for the group instead of the
channel to increase the chance of a cache hit.